+++ tic-tac-toe-app/src/functions/square.tsx
@@ -1,0 +1,12 @@
+interface SquareProps {
+    value: any,
+    onClick:() => any
+}
+
+export function Square(props: SquareProps) {
+    return (
+        <button className="square" onClick={ props.onClick }>
+        { props.value }
+        </button>
+    );
+}
